Extract common accessibility functions
